refactor(common): de-dup `MockAsyncStreamingWriteRpc`
